fix(modal): remove non existant cls-panel option (#1014)
The 'clsPanel' option got renamed to 'selPanel' which indicates the
real usage of it. It's not a css class which will be added for user
specifics, but it is a selector for the modal internally. It won't
set any classes on the modal element and should be removed from the
ember component arguments as well.